relocate installer test
